Refactor base `db` definition into a separate module to break import cycles.
Previously we imported the often needed `db` reference from extensions,
which in turn imported several other modules - creating a fragile import
cycle.

Removed unused imports and ordered as per PEP8 as well where problems
were noticed.
