+++ portal/models/role.py
@@ -10,7 +10,7 @@ To extend the list of roles, add name: description pairs to the
 STATIC_ROLES dict within.
 
 """
-from ..extensions import db
+from ..database import db
 from UserDict import IterableUserDict
