+++ portal/views/user.py
@@ -6,7 +6,8 @@ from sqlalchemy.orm.exc import NoResultFound
 from werkzeug.exceptions import Unauthorized
 
 from ..audit import auditable_event
-from ..extensions import db, oauth, user_manager
+from ..database import db
+from ..extensions import oauth, user_manager
 from ..models.audit import Audit
 from ..models.group import Group
 from ..models.organization import Organization
